+++ R/tibble-package.R
@@ -2,11 +2,14 @@
 #' @importFrom utils head tail
 #' @importFrom pkgconfig set_config
 #' @import rlang
-#' @import lifecycle
 #' @import ellipsis
+#' @importFrom lifecycle deprecate_soft deprecate_warn expect_deprecated
 #' @importFrom vctrs vec_as_location vec_as_location2 vec_as_names vec_as_names_legacy vec_c
-#' @importFrom vctrs vec_is vec_rbind vec_recycle vec_size vec_slice vec_slice<-
+#' @importFrom vctrs vec_is vec_rbind vec_recycle vec_size vec_slice vec_assign
 #' @importFrom vctrs unspecified vec_as_subscript2 num_as_location vec_ptype_abbr
+#' @importFrom vctrs vec_names vec_names2 vec_set_names
+#' @importFrom vctrs new_rcrd
+#' @importFrom vctrs new_data_frame
 #' @aliases NULL tibble-package
 #' @details
 #' `r lifecycle::badge("stable")`
@@ -33,9 +36,5 @@
 #'   * Create a tibble: [tibble()], [as_tibble()], [tribble()], [enframe()]
 #'   * Inspect a tibble: [print.tbl()], [glimpse()]
 #'   * Details on the S3 `tbl_df` class: [`tbl_df-class`]
-#' @inheritSection formatting Package options
-#' @section Package options:
-#' The following option is used for viewing tabular data with `view()`:
-#' - `tibble.view_max`: Maximum number of rows shown if the input is not a
-#'   data frame. Default: 1000.
+#'   * Package options: [tibble_options]
 "_PACKAGE"