+++ R/tibble.R
@@ -31,8 +31,9 @@
 #'   Arguments are evaluated sequentially.
 #'   You can refer to previously created elements directly or using the [.data]
 #'   pronoun.
-#'   An existing `.data` pronoun, provided e.g. inside [dplyr::mutate()],
-#'   is not available.
+#'   To refer explicitly to objects in the calling environment, use [`!!`] or
+#'   [.env], e.g. `!!.data` or `.env$.data` for the special case of an object
+#'   named `.data`.
 #' @param .rows The number of rows, useful to create a 0-column tibble or
 #'   just as an additional check.
 #' @param .name_repair Treatment of problematic column names:
@@ -121,9 +122,9 @@
 #'   )
 #' )
 #' tibble(
-#'   a = 1:4,
-#'   b = diag(4),
-#'   c = cov(iris[1:4])
+#'   a = 1:3,
+#'   b = diag(3),
+#'   c = cor(trees)
 #' )
 #'
 #' # data can not contain POSIXlt columns, or tibbles or matrices
@@ -143,14 +144,20 @@
 #' # You can splice-unquote a list of quosures and expressions:
 #' tibble(!!! list(x = rlang::quo(1:10), y = quote(x * 2)))
 #'
+#' # Use .data, .env and !! to refer explicitly to columns or outside objects
+#' a <- 1
+#' tibble(a = 2, b = a)
+#' tibble(a = 2, b = .data$a)
+#' tibble(a = 2, b = .env$a)
+#' tibble(a = 2, b = !!a)
+#' try(tibble(a = 2, b = .env$bogus))
+#' try(tibble(a = 2, b = !!bogus))
 tibble <- function(...,
                    .rows = NULL,
                    .name_repair = c("check_unique", "unique", "universal", "minimal")) {
   xs <- quos(...)
 
-  is.null <- map_lgl(xs, quo_is_null)
-
-  tibble_quos(xs[!is.null], .rows, .name_repair)
+  tibble_quos(xs, .rows, .name_repair)
 }
 
 #' tibble_row()
@@ -165,14 +172,12 @@ tibble <- function(...,
 #' @examples
 #'
 #' # Use tibble_row() to construct a one-row tibble:
-#' tibble_row(a = 1, lm = lm(Petal.Width ~ Petal.Length + Species, data = iris))
+#' tibble_row(a = 1, lm = lm(Height ~ Girth + Volume, data = trees))
 tibble_row <- function(...,
                        .name_repair = c("check_unique", "unique", "universal", "minimal")) {
-  xs <- quos(...)
+  xs <- enquos(...)
 
-  is.null <- map_lgl(xs, quo_is_null)
-
-  tibble_quos(xs[!is.null], .rows = 1, .name_repair = .name_repair, single_row = TRUE)
+  tibble_quos(xs, .rows = 1, .name_repair = .name_repair, single_row = TRUE)
 }
 
 #' Test if the object is a tibble
@@ -258,6 +263,10 @@ tibble_quos <- function(xs, .rows, .name_repair, single_row = FALSE) {
   }
 
   names(output) <- col_names
+
+  is_null <- map_lgl(output, is.null)
+  output <- output[!is_null]
+
   output <- splice_dfs(output)
   output <- set_repaired_names(output, repair_hint = TRUE, .name_repair = .name_repair)
 
@@ -325,7 +334,7 @@ error_tibble_row_size_one <- function(j, name, size) {
     desc <- paste0("at position ", j)
   }
 
-  tibble_error(bullets(
+  tibble_error(problems(
     "All vectors must be size one, use `list()` to wrap.",
     paste0("Column ", desc, " is of size ", size, ".")
   ))
@@ -350,12 +359,10 @@ error_incompatible_size <- function(.rows, vars, vars_len, rows_source) {
     paste0("Size ", x, ": ", pluralise_commas(text, y))
   })
 
-  problems <- set_default_name(problems, "")
-
-  tibble_error(problems(
+  tibble_error(bullets(
     "Tibble columns must have compatible sizes:",
     if (!is.null(.rows)) paste0("Size ", .rows, ": ", rows_source),
     problems,
-    i = "Only values of size one are recycled"
+    info = "Only values of size one are recycled."
   ))
 }