+++ R/zzz.R
@@ -5,6 +5,11 @@
   } else {
     safe_match <<- safe_match_default
   }
+
+  num <<- pillar::num
+  set_num_opts <<- pillar::set_num_opts
+  char <<- pillar::char
+  set_char_opts <<- pillar::set_char_opts
 }
 
 safe_match_3_0 <- function(x, table) {