+++ tests/testthat/test-tbl_df.R
@@ -3,6 +3,21 @@ test_that("as.data.frame()", {
   expect_identical(as.data.frame(as_tibble(df)), df)
 })
 
+test_that("as.data.frame() strips inner names (#837)", {
+  tbl <- tibble(a = c(b = 1:3))
+  expect_identical(as.data.frame(tbl)$a, 1:3)
+})
+
+test_that("as.data.frame() keeps inner names for lists (#837)", {
+  tbl <- tibble(a = list(b = 1:3))
+  expect_identical(as.data.frame(tbl)$a, list(b = 1:3))
+})
+
+test_that("as.data.frame() keeps inner names for records", {
+  tbl <- tibble(x = new_rcrd(list(a = 1:3)))
+  expect_identical(as.data.frame(tbl)$x, new_rcrd(list(a = 1:3)))
+})
+
 test_that("names<-()", {
   new_tbl <- function(...) {
     data <- list(1, "b")
@@ -50,7 +65,6 @@ test_that("names<-()", {
 
 test_that("output test", {
   skip_if_not_installed("testthat", "3.0.0.9000")
-  skip_if_not_installed("lifecycle", "0.2.0.9000")
 
   expect_snapshot({
     df <- tibble(a = 1, b = 2)