+++ tests/testthat/test-vignette-digits.R
@@ -1,0 +1,3 @@
+test_that("digits vignette", {
+  test_galley("digits")
+})