+++ tests/testthat/test-vignette-types.R
@@ -1,0 +1,3 @@
+test_that("types vignette", {
+  test_galley("types")
+})